requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Werkzeug==3.0.1
nltk==3.8.1
newspaper3k==0.2.8
langdetect==1.0.9
textblob==0.17.1
flask-pymongo>=2.3.0
pymongo>=4.3.0
python-dotenv>=1.0.0
youtube-transcript-api>=0.5.0
gunicorn==21.2.0
yt-dlp
requests
python-dotenv